refactor: move trending products endpoint from ProductItemController to ProductController and update documentation
