+++ src/main/kotlin/org/shangahi/sellio_backend/api/controller/ProductItemController.kt
@@ -4,18 +4,12 @@ import io.swagger.v3.oas.annotations.tags.Tag
 import jakarta.validation.Valid
 import org.shangahi.sellio_backend.api.dto.request.ProductItemRequest
 import org.shangahi.sellio_backend.api.dto.request.ProductItemUpdateRequest
-import org.shangahi.sellio_backend.api.dto.response.PageResponse
 import org.shangahi.sellio_backend.api.dto.response.ProductItemResponse
-import org.shangahi.sellio_backend.api.dto.response.TrendingProductResponse
 import org.shangahi.sellio_backend.api.mapper.toResponse
 import org.shangahi.sellio_backend.api.swagger.doc.ProductItemDoc
 import org.shangahi.sellio_backend.service.ProductItemService
-import org.springdoc.core.annotations.ParameterObject
-import org.springframework.data.domain.Pageable
-import org.springframework.data.web.PageableDefault
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
-import org.springframework.security.core.annotation.AuthenticationPrincipal
 import org.springframework.web.bind.annotation.*
 import java.util.*
 
@@ -32,18 +26,6 @@ class ProductItemController(
     ): List<ProductItemResponse> {
         val productItems = productItemService.getProductItems(productId)
         return productItems.map { it.toResponse() }
-    }
-
-    @ProductItemDoc.GetTrendingProducts
-    @GetMapping("/trending")
-    fun getTrendingProducts(
-        @AuthenticationPrincipal userId: UUID?,
-        @ParameterObject
-        @PageableDefault(page = 0, size = 20)
-        pageable: Pageable
-    ): PageResponse<TrendingProductResponse> {
-        val trendingProducts = productItemService.getTrendingProducts(userId, pageable)
-        return trendingProducts
     }
 
     @ProductItemDoc.InsertProductItem
